requirements.txt
Flask
opencv-python
numpy
easyocr
pdf2image
fuzzywuzzy
python-Levenshtein
Pillow
scikit-image
google-generativeai
